강의 관련 메서드 전체적으로 수정 + 월별/연도별 그래프 데이터 조회 기능 추가

+++ LearnsMate/src/main/java/intbyte4/learnsmate/contractprocess/repository/ContractProcessRepository.java
@@ -16,5 +16,5 @@ public interface ContractProcessRepository extends JpaRepository<ContractProcess
 
     Optional<ContractProcess> findByLectureAndApprovalProcess(Lecture lecture, Integer approvalProcess);
 
-    long countByLecture(Lecture lecture);
+    int countByLecture(Lecture lecture);
 }
